fix: Fn::RefNow bug fixes
